% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta-package.R
\docType{package}
\name{deltaq}
\alias{deltaq}
\alias{deltaq-package}
\title{deltaq: Shiny Exploration of Quarantine Breach Data}
\description{
The deltaq package provides:
\itemize{
\item The dataset \code{\link{breach}}, containing daily breach rate observations
for Australia and its states/territories.
\item A Shiny app launched via \code{\link{launch_app}} to explore breach rates
interactively with date and state filters.
}
}
\seealso{
\code{\link{launch_app}}, \code{\link{breach}}
}
\author{
\strong{Maintainer}: Ibtesam Ifaz \email{iifa0001@student.monash.edu}

}
\keyword{internal}
