% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breach.R
\docType{data}
\name{breach}
\alias{breach}
\title{Breach Data (sheet 9 of covid.xlsx)}
\format{
A tibble with one row per date and several numeric columns.
}
\source{
Local Excel file \code{data-raw/covid.xlsx}
}
\usage{
breach
}
\description{
Data extracted from sheet 9 of the Excel file stored in \code{data-raw/covid.xlsx}.
Includes daily observations of quarantine breaches and related metrics by state.
}
\examples{
data(breach)
head(breach)
}
\keyword{datasets}
